Update lab0 to python 3 and big lab0 updates.
Lab0 has been updated to python 3, as well as all
of the other labs (Thanks nathan!).
In addition, it has been decided that we'll release
lab0 in early august to give students a chance to
work on their python and as a diagnostic for them
to figure out if their python needs improving before
the class starts. Included in this commit, and
the special lab0 august release, is a modified
tester.py which only supports offline testing.
Online testing is not available during this "pre-release."